+++ docs/source/setup/quickstart.rst
@@ -50,8 +50,8 @@ Training outputs, including checkpoints, are saved under ``logs/``. Add
    For example, ``--extra ovphysx`` makes the OV PhysX integration available, while
    ``physics=ovphysx`` selects it for the task. You can combine extras as needed. The ``--extra all``
    shortcut installs a curated set of backends, RL libraries, and visualizers.
-   Specialized extras such as ``rlinf``, ``mimic``, ``teleop``, ``tetrahedralization``, ``video``,
-   and ``leapp`` are not included; add them explicitly when needed. See
+   Specialized extras such as ``ray``, ``rlinf``, ``mimic``, ``teleop``, ``tetrahedralization``,
+   ``video``, and ``leapp`` are not included; add them explicitly when needed. See
    :ref:`installation-optional-extras` for the complete list.
 
 Choose an RL library
