+++ source/isaaclab_contrib/isaaclab_contrib/rl/ray/__init__.py
@@ -3,8 +3,4 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
-"""Deprecated wrapper for :mod:`isaaclab_contrib.rl.ray.launch`."""
-
-from _compat import forward
-
-forward("isaaclab_contrib.rl.ray.launch", globals())
+"""Contributed Ray job-dispatch and hyperparameter-tuning workflows."""
